requirements.txt
absl-py==1.2.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrdict==2.0.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
bce-python-sdk==0.8.79
beautifulsoup4==4.11.1
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.7
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
cython-bbox==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.6
debugpy-run==1.6
decorator==5.1.1
deep-sort-realtime==1.3.1
dill==0.3.6
dlib==19.24.0
easydict==1.10
einops==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
face-alignment==1.3.5
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.92.0
ffmpeg==1.4
ffmpy==0.3.0
filelock==3.8.0
filterpy==1.4.5
fire==0.5.0
flake8==5.0.4
Flask==2.2.3
flask-babel==3.0.1
flatbuffers==23.3.3
fonttools==4.37.3
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.2
fvcore==0.1.5.post20220512
gdown==4.5.1
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.12.0
google-auth-oauthlib==0.4.6
gradio==3.19.1
greenlet==2.0.2
grpcio==1.49.1
h11==0.14.0
h5py==3.7.0
httpcore==0.16.3
httpx==0.23.3
imageio==2.22.0
imgaug==0.4.0
importlib-metadata==4.2.0
importlib-resources==5.12.0
imutils==0.5.4
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.4
lap==0.4.0
linkify-it-py==1.0.3
llvmlite==0.39.1
lmdb==1.4.0
loguru==0.6.0
lxml==4.9.2
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.9.0.1
mkl-fft==1.3.1
mkl-service==2.4.0
motmetrics==1.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==2.6.3
ninja==1.11.1
numba==0.56.4
oauthlib==3.2.1
onemetric==0.1.2
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.8
onnxruntime==1.8.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.1
opt-einsum==3.3.0
orjson==3.8.6
packaging==21.3
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==1.3.5
parso==0.8.3
pdf2docx==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
portalocker==2.5.1
premailer==3.10.0
prompt-toolkit==3.0.31
protobuf==3.20.0
psutil==5.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycodestyle==2.9.1
pycryptodome==3.17
pydantic==1.10.5
pydub==0.25.1
pyflakes==2.5.0
Pygments==2.13.0
PyMuPDF==1.20.2
PyOpenGL==3.1.6
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.5
python-rapidjson==1.9
pytorch3d==0.3.0
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
rapidfuzz==2.13.7
rarfile==4.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.2
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
seaborn==0.12.0
sentry-sdk==1.16.0
shapely==2.0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.25.0
supervision==0.1.0
sympy==1.10.1
tabulate==0.8.10
tb-nightly==2.11.0a20221004
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
thop==0.1.1.post2209072238
tifffile==2021.11.2
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchreid==0.2.5
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trimesh==3.15.2
tritonclient==2.30.0
typing-inspect==0.8.0
uc-micro-py==1.0.1
ultralytics==8.0.48
uvicorn==0.20.0
visualdl==2.5.0
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
x2paddle==1.4.0
xmltodict==0.13.0
yacs==0.1.8
yapf==0.32.0
yarl==1.8.2
zipp==3.8.1
zope.event==4.6
zope.interface==5.5.2
